Eliminate util/ and utils/ directories - consolidate or inline all utilities
Complete audit and consolidation of utility directories, removing 4,029 lines
across 18 files by eliminating unused abstractions and inlining single-use code.

## Deletions

### util/ directory (12 files, 3,346 lines):
**UNUSED - Deleted (9 files, 2,257 lines):**
- config_manager.py (268 lines) - 0 imports
- config_presets.py (248 lines) - 0 imports
- env_loader.py (112 lines) - 0 imports
- error_handling.py (355 lines) - 0 imports
- logging_config.py (55 lines) - 0 imports
- path_security.py (230 lines) - 0 imports
- rate_limiter.py (128 lines) - 0 imports
- rich_console.py (417 lines) - 0 imports
- system_detection.py (244 lines) - 0 imports

**INLINED (2 files, 909 lines):**
- input_validation.py (638 lines) → validators.py (added 34 lines)
  - Inlined InputValidationError class and validate_coordinates function
- invalid_data_tracker.py (271 lines) → distance/engine.py
  - Removed unused tracking call (16 lines deleted)

**DELETED:**
- __init__.py (380 lines) - exported above modules

### utils/ directory (2 files, 92 lines):
**DUPLICATE - Deleted:**
- api_key.py (92 lines) - Duplicate of security/utils.py
  - Only used in tests, production code uses security/utils
- __init__.py (0 lines) - empty

### Test files (2 files, 569 lines):
- test_api_key_utility.py (314 lines) - tested deleted utils/api_key.py
- test_api_key_integration.py (255 lines) - tested deleted utils/api_key.py

## Impact

**Total removed**: 4,029 lines across 18 files
**Directories removed**: 2 (util/, utils/)
**Net change**: +34 lines added to validators.py, -4,063 lines total

## Benefits

✅ Eliminates "just put it in utils" anti-pattern
✅ Functions moved closer to their usage
✅ Removes unnecessary abstractions
✅ Clearer code organization
✅ No functionality lost - unused code or duplicates only

## Verification

✓ All validator tests pass (22/22)
✓ Main imports work correctly
✓ Inlined code functions identically

Fixes #112
Part of #100 (systematic simplification)
